issue/#171-refactor-file-structure * refactor backend file structure and update imports
