+++ backend/src/test/java/backend/unit/GameServiceUnitTests.java
@@ -12,8 +12,8 @@ import backend.domain.Language;
 import backend.domain.dto.FixedLetterResponse;
 import backend.domain.dto.GameGridRequest;
 import backend.domain.dto.ValidationResultResponse;
-import backend.domain.entities.FinnishWord;
-import backend.domain.entities.Word;
+import backend.domain.entity.FinnishWord;
+import backend.domain.entity.Word;
 import backend.service.GameService;
 import backend.service.RepositoryService;
 import backend.service.UtilityService;
